Put the Lid-on and walk away 28th Oct 2017
GPS/OCXO displine - working
Si5351a Synthesizer running from the OCXO - working
AD9910 DDS module - abandoned

+++ Freqref/ocxo.h
@@ -32,7 +32,7 @@ void resetcnt(int);
 // proportional control
 void propocxo();
 
-void setdacandwait(int,int);
+void setdacandwait(uint16_t,uint16_t);
 
 /*
 Read the 4 bytes = 32 bits out of each counter
